Printing of RSNA ID and Confirmation of Patient
When selecting a patient a dialog prompts the user to confirm that this
is the correct patient and that the patient has consented to the program
before showing exams for that patient.

After creating an RSNA id a new window will open in addition to the
exams list for printing of that RSNA id.  This can also be launched by a
primary link.