Added Consent Check
The app now checks to see if a consent duration key has been given in
the configurations table (defaulting to 1 year).  If they have not
consented within the given duration the user is prompted with a consent
dialog box.  When consent is confirmed in the dialog box the timestamp
is recorded in the database for that patient (patientsconsent_timestamp)