RSNA ID and Immediate send logic
The application now gives the user the option to use an email based rsna
id as part of the single use patient id instead of a token. It also
provides a third option to the sending where it will send images without
waiting for a final report. The language for the sending options has
also been updated.